requirements.txt
tqdm==4.66.1
python-dotenv==1.0.0

pandas==2.1.1
numpy==1.26.1
matplotlib==3.8.0
scikit-learn==1.3.1
jupyter==1.0.0

### development
black==23.10.0
isort==5.12.0
bandit==1.7.5
mypy==1.6.1
pylint==3.0.2

### add specific project dependencies here:
seaborn==0.13.0
torch==2.1.0
torchvision==0.16.0
grad-cam==1.4.8
opencv-python==4.8.1.78
